// generated/openapi/model/api-error.ts
/**
 * Workflow Approval System API
 * API for a SaaS platform that allows customers to manage approvals for generic workflows, users, and groups.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { APIErrorDetailsInner } from './api-error-details-inner';


export interface APIError { 
    /**
     * Error code
     */
    code?: string;
    /**
     * Error message
     */
    message?: string;
    details?: Array<APIErrorDetailsInner>;
}